requirements.txt
Flask==3.0.3
inference_sdk==0.23.0
numpy==2.1.2
opencv_python==4.10.0.84
Pillow==11.0.0
streamlit==1.37.0
ultralytics==8.3.13
